Management

+++ src/Employee/EmployeeWorkingTime.java
@@ -21,13 +21,13 @@ public class EmployeeWorkingTime implements TableEntry{
 	private Employee employee;
 	private LocalDateTime punchIn;
 	private LocalDateTime punchOut;
-	private double workingHour;
+	private long workingHour; // compute by minutes
 	
 	public EmployeeWorkingTime() {
 
 	}
 	public EmployeeWorkingTime(long workingTimeID, Employee employee, LocalDateTime punchIn, LocalDateTime punchOut,
-			double workingHour) {
+			long workingHour) {
 		super();
 		this.workingTimeID = workingTimeID;
 		this.employee = employee;
@@ -59,10 +59,10 @@ public class EmployeeWorkingTime implements TableEntry{
 	public void setPunchOut(LocalDateTime punchOut) {
 		this.punchOut = punchOut;
 	}
-	public double getWorkingHour() {
+	public long getWorkingHour() {
 		return workingHour;
 	}
-	public void setWorkingHour(double workingHour) {
+	public void setWorkingHour(long workingHour) {
 		this.workingHour = workingHour;
 	}
